:arrow_up: Add Askja example Figure 10
Add Figure 10 script (as Jupyter Notebook). Note that to
generate this figure requires (a) a tweaked QM run that outputs
both the marginalised coalescence map (requires source-code
modification) and (b) the cut waveforms (just an edit to the
locate.py script).

The added i/o functions for handling marginalised coalescence maps
are also used within the figure script to read in the saved map.