:wrench: Update Figs 9&10
Figure code is broadly unchanged, other than some minor backslash syntax
changes apparently enforced in Python 3.12, and a move to specifying the
event to be plotted as a user argument in Figure 10, so it's easy to
switch between events (all having cut waveforms and marginal coalescence
maps output now, as of the previous commit).

Main change is using the updated results. Slightly fewer events are used
in Figure 9 now (39 vs 43 previously), but this is just playing around
the noise level so I've left it at a slightly higher value to keep
things clean and simple. Note the trigger smoothing definitely does help
things in this case. Still a slight question in my mind whether to stick
with this event for Fig10, but easy to change.