:wrench: Update synthetic scripts
Update synthetic scripts to reflect updates to QuakeMigrate + a general
cleanup of the simulate module.

Note: This will currently NOT produce all the required outputs for the
synthetic example figures, as we need to add the ability to write out
the 4-D coalescence maps. I think this should be straightforward to get
done and included with v1.2.0, so everything will run out of the box as
intended.