:electron: add icequake example
Add code to run the Rutford icequake example presented in the QuakeMigrate
manuscript. Also add the scripts used to build Figures 7 and 8, which
show a snippet of the trigger output and the final located catalogue on
a map, respectively.

Note: The script to create Figure 7 will not run with QuakeMigrate in
its current state (v1.2.0). This is because it also requires the minimum
and maximum times of the triggered event window, which is not written to
file by default. A toggle was added locally for the purpose of making
the figure for the manuscript, which should be merged with the core
package at some point.