:atom: add QuakeMigrate profiling results
Add a set of scripts that produce the runtime and memory usage profiling
done for the supplementary material in the QuakeMigrate manuscript.

The profiling is based on the Askja VT-DLP example presented in the main
manuscript. A single hour is chosen and a parameter search is made over
the 3 primary controls on runtime/memory usage:

- `n_threads`: the number of CPU threads to be used for parallel
  processing
- `decimation_factor`: the decimation factor used for the traveltime LUT
- `timestep`: the amount of waveform data processed at a time

The results of this parameter search are plotted in Supplementary Figure
3, with a focus on the peak memory use (the memory high watermark) and
the overall runtime.

Following this parameter search, each stage of QM is profiled to show
the memory usage as a function of time.