:bug: specify correctly the 1800 s timestep
Specify correctly the 1800 s timestep run, which had not been properly
changed from 2400 s. 2400 s (that is, 40 minutes) did not divide into an
hour (3600 s), so QuakeMigrate was automatically adjusting the endtime
to incorporate the full time period requested by extending for an
additional 1200 s. This was causing an apparent increase in computation
time as a function of timestep size, though we expect it to be
essentially flat with increasing timesteps.