:wrench: adjust dimension and colouring
Adjust figure dimension to use cm, as specified by journal.

Also slightly change how the grid search is represented, with the
currently computed element indicated by the teal square, while the
elements that have been 'searched' are now in grey. Hopefully makes it
more clear that the middle and right-hand panels correspond to the
migration and stack for an earthquake with origin on that square.