:wrench: update docstrings and get data script
Update docstrings to provide manuscript information.

Also adjust the icequake get data script to write the waveform data to
correct day files in the archive. Previously, this required re-reading
and writing of the waveform data. While a little uglier, I think the new
method of downloading the parts required from each day in the example
requires less effort on the part of the user to run.