:arrow_up: Add Askja event P-only location supp fig
Add a new supplementary figure showing a comparison between P-only and
full (P- and S-phase) location for an event from the Askja example.
Scripts to run trigger and locate with P-only are added, (similar to the
setup used for the synthetic supp figure after which this is modelled),
as well as the .afdesign file where the two locate summary plots are
combined (and mildly edited - to make linewidths more visible).